requirements.txt
Pillow
redis
secrets
json
tkinter
